rename `commandoutput` file to `commandio`
